wallet: move deprecated methods to `deprecated.go`
