+++ wallet/recovery.go
@@ -8,195 +8,11 @@ import (
 	"github.com/btcsuite/btcd/btcutil"
 	"github.com/btcsuite/btcd/btcutil/hdkeychain"
 	"github.com/btcsuite/btcd/chaincfg"
-	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcd/wire"
 	"github.com/btcsuite/btcwallet/waddrmgr"
-	"github.com/btcsuite/btcwallet/walletdb"
 	"github.com/btcsuite/btcwallet/wtxmgr"
 )
-
-// RecoveryManager maintains the state required to recover previously used
-// addresses, and coordinates batched processing of the blocks to search.
-//
-// TODO(yy): Deprecated, remove.
-type RecoveryManager struct {
-	// recoveryWindow defines the key-derivation lookahead used when
-	// attempting to recover the set of used addresses.
-	recoveryWindow uint32
-
-	// started is true after the first block has been added to the batch.
-	started bool
-
-	// blockBatch contains a list of blocks that have not yet been searched
-	// for recovered addresses.
-	blockBatch []wtxmgr.BlockMeta
-
-	// state encapsulates and allocates the necessary recovery state for all
-	// key scopes and subsidiary derivation paths.
-	state *RecoveryState
-
-	// chainParams are the parameters that describe the chain we're trying
-	// to recover funds on.
-	chainParams *chaincfg.Params
-}
-
-// NewRecoveryManager initializes a new RecoveryManager with a derivation
-// look-ahead of `recoveryWindow` child indexes, and pre-allocates a backing
-// array for `batchSize` blocks to scan at once.
-//
-// TODO(yy): Deprecated, remove.
-func NewRecoveryManager(recoveryWindow, batchSize uint32,
-	chainParams *chaincfg.Params) *RecoveryManager {
-
-	return &RecoveryManager{
-		recoveryWindow: recoveryWindow,
-		blockBatch:     make([]wtxmgr.BlockMeta, 0, batchSize),
-		chainParams:    chainParams,
-		state: NewRecoveryState(
-			recoveryWindow, chainParams, nil,
-		),
-	}
-}
-
-// Resurrect restores all known addresses for the provided scopes that can be
-// found in the walletdb namespace, in addition to restoring all outpoints that
-// have been previously found. This method ensures that the recovery state's
-// horizons properly start from the last found address of a prior recovery
-// attempt.
-//
-// TODO(yy): Deprecated, remove.
-func (rm *RecoveryManager) Resurrect(ns walletdb.ReadBucket,
-	scopedMgrs map[waddrmgr.KeyScope]waddrmgr.AccountStore,
-	credits []wtxmgr.Credit) error {
-
-	// First, for each scope that we are recovering, rederive all of the
-	// addresses up to the last found address known to each branch.
-	for keyScope, scopedMgr := range scopedMgrs {
-		// Load the current account properties for this scope, using the
-		// the default account number.
-		// TODO(conner): rescan for all created accounts if we allow
-		// users to use non-default address
-		scopeState := rm.state.StateForScope(keyScope)
-		acctProperties, err := scopedMgr.AccountProperties(
-			ns, waddrmgr.DefaultAccountNum,
-		)
-		if err != nil {
-			return err
-		}
-
-		// Fetch the external key count, which bounds the indexes we
-		// will need to rederive.
-		externalCount := acctProperties.ExternalKeyCount
-
-		// Walk through all indexes through the last external key,
-		// deriving each address and adding it to the external branch
-		// recovery state's set of addresses to look for.
-		for i := uint32(0); i < externalCount; i++ {
-			keyPath := externalKeyPath(i)
-			addr, err := scopedMgr.DeriveFromKeyPath(ns, keyPath)
-			if err != nil && err != hdkeychain.ErrInvalidChild {
-				return err
-			} else if err == hdkeychain.ErrInvalidChild {
-				scopeState.ExternalBranch.MarkInvalidChild(i)
-				continue
-			}
-
-			scopeState.ExternalBranch.AddAddr(i, addr.Address())
-		}
-
-		// Fetch the internal key count, which bounds the indexes we
-		// will need to rederive.
-		internalCount := acctProperties.InternalKeyCount
-
-		// Walk through all indexes through the last internal key,
-		// deriving each address and adding it to the internal branch
-		// recovery state's set of addresses to look for.
-		for i := uint32(0); i < internalCount; i++ {
-			keyPath := internalKeyPath(i)
-			addr, err := scopedMgr.DeriveFromKeyPath(ns, keyPath)
-			if err != nil && err != hdkeychain.ErrInvalidChild {
-				return err
-			} else if err == hdkeychain.ErrInvalidChild {
-				scopeState.InternalBranch.MarkInvalidChild(i)
-				continue
-			}
-
-			scopeState.InternalBranch.AddAddr(i, addr.Address())
-		}
-
-		// The key counts will point to the next key that can be
-		// derived, so we subtract one to point to last known key. If
-		// the key count is zero, then no addresses have been found.
-		if externalCount > 0 {
-			scopeState.ExternalBranch.ReportFound(externalCount - 1)
-		}
-		if internalCount > 0 {
-			scopeState.InternalBranch.ReportFound(internalCount - 1)
-		}
-	}
-
-	// In addition, we will re-add any outpoints that are known the wallet
-	// to our global set of watched outpoints, so that we can watch them for
-	// spends.
-	for _, credit := range credits {
-		_, addrs, _, err := txscript.ExtractPkScriptAddrs(
-			credit.PkScript, rm.chainParams,
-		)
-		if err != nil {
-			return err
-		}
-
-		rm.state.AddWatchedOutPoint(&credit.OutPoint, addrs[0])
-	}
-
-	return nil
-}
-
-// AddToBlockBatch appends the block information, consisting of hash and height,
-// to the batch of blocks to be searched.
-//
-// TODO(yy): Deprecated, remove.
-func (rm *RecoveryManager) AddToBlockBatch(hash *chainhash.Hash, height int32,
-	timestamp time.Time) {
-
-	if !rm.started {
-		log.Infof("Seed birthday surpassed, starting recovery "+
-			"of wallet from height=%d hash=%v with "+
-			"recovery-window=%d", height, *hash, rm.recoveryWindow)
-		rm.started = true
-	}
-
-	block := wtxmgr.BlockMeta{
-		Block: wtxmgr.Block{
-			Hash:   *hash,
-			Height: height,
-		},
-		Time: timestamp,
-	}
-	rm.blockBatch = append(rm.blockBatch, block)
-}
-
-// BlockBatch returns a buffer of blocks that have not yet been searched.
-//
-// TODO(yy): Deprecated, remove.
-func (rm *RecoveryManager) BlockBatch() []wtxmgr.BlockMeta {
-	return rm.blockBatch
-}
-
-// ResetBlockBatch resets the internal block buffer to conserve memory.
-//
-// TODO(yy): Deprecated, remove.
-func (rm *RecoveryManager) ResetBlockBatch() {
-	rm.blockBatch = rm.blockBatch[:0]
-}
-
-// State returns the current RecoveryState.
-//
-// TODO(yy): Deprecated, remove.
-func (rm *RecoveryManager) State() *RecoveryState {
-	return rm.state
-}
 
 // RecoveryState manages the initialization and lookup of ScopeRecoveryStates
 // for any actively used key scopes.
@@ -758,32 +574,6 @@ func (rs *RecoveryState) expandHorizons() (bool, error) {
 	}
 
 	return expanded, nil
-}
-
-// ScopeRecoveryState is used to manage the recovery of addresses generated
-// under a particular BIP32 account. Each account tracks both an external and
-// internal branch recovery state, both of which use the same recovery window.
-//
-// TODO(yy): Deprecated, remove.
-type ScopeRecoveryState struct {
-	// ExternalBranch is the recovery state of addresses generated for
-	// external use, i.e. receiving addresses.
-	ExternalBranch *BranchRecoveryState
-
-	// InternalBranch is the recovery state of addresses generated for
-	// internal use, i.e. change addresses.
-	InternalBranch *BranchRecoveryState
-}
-
-// NewScopeRecoveryState initializes an ScopeRecoveryState with the chosen
-// recovery window.
-//
-// TODO(yy): Deprecated, remove.
-func NewScopeRecoveryState(recoveryWindow uint32) *ScopeRecoveryState {
-	return &ScopeRecoveryState{
-		ExternalBranch: NewBranchRecoveryState(recoveryWindow, nil),
-		InternalBranch: NewBranchRecoveryState(recoveryWindow, nil),
-	}
 }
 
 // BranchRecoveryState maintains the required state in-order to properly
